Рефакторинг cli части (плавная миграция на typer): вынос опций в options/*, использование typer
Typer добавляет визуальные разделения в справке, можно удобно управлять help сообщениями из самих объектов опций.
