+++ ymd/cli/filename_regexps.py
@@ -1,0 +1,7 @@
+
+import re
+
+TRACK_RE = re.compile(r"track/(\d+)")
+ALBUM_RE = re.compile(r"album/(\d+)$")
+ARTIST_RE = re.compile(r"artist/(\d+)$")
+PLAYLIST_RE = re.compile(r"([\w\-._@]+)/playlists/(\d+)$")
